+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## (Unreleased)
 
 ### Monument
+- (#91) Calls can now go from/to different lead labels.  Set this with e.g.
+    `lead_location = { from = "2nds", to = "HL" }`.  Useful for adding finer control over where
+    calls can be placed.
+- (#91) Allow multiple labels on the same row within a lead.  Also reversed the syntax from e.g.
+    `lead_locations = { 0 = "LE", 16 = "HL" }` to `lead_locations = { LE = 0, HL = 16 }`.  The same
+    label can be added to multiple rows like `lead_locations = { SE = [3, 9] }` (for Six-Ends in
+    Stedman).
 - (#89) Refactor the search algorithm (splitting the node expansion from the best-first search code)
 
 ---
@@ -50,6 +57,9 @@
 
 ### Internal Improvements
 - (#64) Use `goldilocks-json-fmt` to format the test result files.
+
+### BellFrame
+- (#??) Allow multiple lead labels to be placed on the same row
 
 ---